# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scholiax API",
    "description": "Backend API for Scholiax - Academic communication tool for lecturers and students",
    "version": "1.0.0",
    "contact": {
      "name": "Scholiax Team"
    }
  },
  "servers": [
    {
      "url": "https://scholiax-1.onrender.com/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["student", "lecturer"] },
          "enrolledCourses": { "type": "array", "items": { "type": "string" } },
          "createdCourses": { "type": "array", "items": { "type": "string" } },
          "avatar": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "code": { "type": "string" },
          "description": { "type": "string" },
          "lecturer": { "$ref": "#/components/schemas/User" },
          "students": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "body": { "type": "string" },
          "course": { "$ref": "#/components/schemas/Course" },
          "createdBy": { "$ref": "#/components/schemas/User" },
          "isImportant": { "type": "boolean" },
          "attachments": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "string" } }
        }
      }
    }
  },
  "paths": {
...
    "/courses": {
      "post": {
        "tags": ["Courses"],
        "summary": "Create a new course",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "code", "description"],
                "properties": {
                  "title": { "type": "string" },
                  "code": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Course" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/Error" },
          "403": { "$ref": "#/components/schemas/Error" }
        }
      },
      "get": {
        "tags": ["Courses"],
        "summary": "Get all courses",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 10 },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "description": "Search term"
          }
        ],
        "responses": {
          "200": {
            "description": "List of courses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Course" }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "total": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/announcements": {
      "post": {
        "tags": ["Announcements"],
        "summary": "Create announcement",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "body", "course"],
                "properties": {
                  "title": { "type": "string" },
                  "body": { "type": "string" },
                  "course": { "type": "string" },
                  "isImportant": { "type": "boolean" },
                  "attachments": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Announcement created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Announcement" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/schemas/Error" },
          "403": { "$ref": "#/components/schemas/Error" }
        }
      }
    },
...
  }
} 